Mother of all errors.
So, while I was being lazy, I simply called sorted(x.path()), not wanting to do a proper "pointer-based" solution, so of course this comes back to bite me.

I had forgotten that I hadn't put comparison operators on this more "basic" nodes, so they technically shouldn't compare anything at all. I was -also- being lazy and only using python2 instead of python3. Again, comes back to bite.

Gist is that comparison shouldn't have worked, but due to the obscure way ptyhon compares arbitrary objects, it did, -sometimes-. It compares the object dictionaries, which themselves have some weird comparison convention. It worked when run in terminal, but not Textmate, so I figured it might be a textmate bug.

I was clued in when adding a method changed the output of "sorted." I guess this changed the class dictionary for Node, messing up the "lucky" sorting order. Long story short, Monkey-Patch with attrgetter